// src/app/model/newCloud.ts
/**
 * Cloudiator REST Api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.2.0
 * Contact: daniel.baur@uni-ulm.de
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Api } from './api';
import { CloudConfiguration } from './cloudConfiguration';
import { CloudCredential } from './cloudCredential';
import { CloudType } from './cloudType';


/**
 *  Represents a new cloud that is to be created 
 */
export interface NewCloud { 
    /**
     * URI where the api of this cloud provider can be accessed.
     */
    endpoint?: string;
    cloudType: CloudType;
    api: Api;
    credential: CloudCredential;
    cloudConfiguration?: CloudConfiguration;
}